Split the Helper class into separate classes
