+++ src/test/java/endpoints/booking/CreateBookingTest.java
@@ -9,12 +9,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static helpers.ClassToJsonConverter.convertClassToJsonWithExtraField;
-import static specs.BaseSpec.requestSpec;
-import static specs.BaseSpec.responseSpec;
 import static io.restassured.RestAssured.given;
 import static io.restassured.module.jsv.JsonSchemaValidator.matchesJsonSchemaInClasspath;
 import static org.hamcrest.Matchers.*;
-import static org.hamcrest.Matchers.equalTo;
+import static specs.BaseSpec.requestSpec;
+import static specs.BaseSpec.responseSpec;
 
 public class CreateBookingTest extends BaseTest {
