+++ src/test/java/endpoints/booking/GetBookingIdsTest.java
@@ -7,7 +7,7 @@ import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static helpers.Helpers.createBooking;
+import static helpers.BookingHelper.createBooking;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.*;
 import static specs.BaseSpec.requestSpec;
